+++ click_extra/commands.py
@@ -38,7 +38,7 @@ from .config import (
     PREPEND_SUBCOMMANDS_KEY,
     ConfigOption,
     ConfigValidator,
-    DumpConfigOption,
+    ExportConfigOption,
     NoConfigOption,
     ValidateConfigOption,
     make_schema_callable,
@@ -86,7 +86,7 @@ def default_params() -> list[click.Option]:
             behavior and value of the other options.
     #. ``--no-config``
     #. ``--validate-config CONFIG_PATH``
-    #. ``--dump-config FORMAT``
+    #. ``--export-config FORMAT``
     #. ``--accessible``
         .. hint::
             ``--accessible`` is placed before ``--color`` and ``--table-format`` so it
@@ -129,7 +129,7 @@ def default_params() -> list[click.Option]:
         ConfigOption(),
         NoConfigOption(),
         ValidateConfigOption(),
-        DumpConfigOption(),
+        ExportConfigOption(),
         AccessibleOption(),
         ColorOption(),
         NoColorOption(),